+++ R/locatome_to_df.R
@@ -1,24 +1,16 @@
 #' Convert a multi-line locatome to a tidy data frame
 #' @inheritParams default_params_doc
+#' @examples
+#' topology_text <- c(">Protein A", "iiiimmmmmooooo")
+#' fasta_text_to_tibble(topology_text)
 #' @export
-locatome_to_df <- function(locatome) {
-
-  n_names <- sum(stringr::str_detect(string = locatome, pattern = "^>"))
-  df <- tibble::tibble(name = rep("", n_names), topology = "")
-
-  # Put into table, row by row
-  cur_row_index <- 0 # Line to write in data frame
-  for (line in locatome) {
-    if (stringr::str_detect(string = line, pattern = "^>")) {
-      cur_row_index <- cur_row_index + 1
-      df$name[cur_row_index] <- line
-    } else {
-      df$topology[cur_row_index] <- paste0(df$topology[cur_row_index], line)
-    }
+locatome_to_df <- function(topology_text, locatome = character(0)) {
+  if  (length(locatome) != 0) {
+    warning(
+      "The argument name 'locatome' is deprecated, use 'topology_text' instead"
+    )
+    topology_text <- locatome
   }
-
-  # Strip the '>' from the name
-  df$name <- stringr::str_replace(df$name, pattern = ">", replacement = "")
-
-  df
+  warning("'locatome_to_df' is deprecated, use 'fasta_text_to_tibble' instead")
+  tmhmm::fasta_text_to_tibble(fasta_text = topology_text)
 }