+++ R/parse_fasta_file.R
@@ -9,5 +9,5 @@ parse_fasta_file <- function(fasta_filename) {
     "'parse_fasta_file' is deprecated, ",
     "use 'pureseqtmr::load_fasta_file_as_tibble_cpp' instead"
   )
-  pureseqtmr::load_fasta_file_as_tibble_cpp(fasta_filename)
+  pureseqtmr::load_fasta_file_as_tibble(fasta_filename)
 }