+++ R/run_tmhmm.R
@@ -3,14 +3,14 @@
 #' The function \link{run_tmhmm_on_one_protein_fasta_file}
 #' is the function that actually calls TMHMM, on a per-protein basis.
 #' @inheritParams default_params_doc
-#' @return text similar to a FASTA file, yet with the
-#'   predicted topology instead of the protein sequence
+#' @return the topology as text, in a similar form as a FASTA file,
+#'   for example \code{c(">Protein A", "iiiimmmmmoooo")}
 #' @author Richèl J.C. Bilderbeek
 #' @examples
 #' if (is_tmhmm_installed()) {
 #'   fasta_filename <- system.file("extdata", "tmhmm.fasta", package = "tmhmm")
-#'   locatome <- run_tmhmm(fasta_filename)
-#'   message(locatome, sep = "\n")
+#'   topology_text <- run_tmhmm(fasta_filename)
+#'   cat(topology_text, sep = "\n")
 #' }
 #' @export
 run_tmhmm <- function(
@@ -21,7 +21,7 @@ run_tmhmm <- function(
   tmhmm::check_tmhmm_installation(folder_name = folder_name)
 
   # All sequences
-  t <- pureseqtmr::load_fasta_file_as_tibble_cpp(fasta_filename)
+  t <- pureseqtmr::load_fasta_file_as_tibble(fasta_filename)
 
   # List of all texts
   lists <- list()
@@ -33,7 +33,7 @@ run_tmhmm <- function(
       t = t[i, ],
       fasta_filename = one_protein_fasta_filename
     )
-    text <- run_tmhmm_on_one_protein_fasta_file(
+    text <- tmhmm::run_tmhmm_on_one_protein_fasta_file(
       one_protein_fasta_filename = one_protein_fasta_filename,
       folder_name = folder_name
     )