+++ R/tally_tmhs.R
@@ -5,7 +5,7 @@
 #' @examples
 #' if (is_tmhmm_installed()) {
 #'   tally_tmhs(
-#'     locatome_to_df(
+#'     fasta_text_to_tibble(
 #'       c(
 #'         ">my_protein",
 #'         "FASTASTIC"