+++ tests/testthat/test-plot_locatome.R
@@ -1,17 +1,7 @@
 test_that("use", {
-  if (!is_tmhmm_installed()) return()
-
-  fasta_filename <- system.file("extdata", "tmhmm.fasta", package = "tmhmm")
-  locatome <- run_tmhmm(fasta_filename)
-  expect_silent(plot_locatome(locatome))
-})
-
-test_that("use", {
-  if (!is_tmhmm_installed()) return()
-
-  fasta_filename <- system.file(
-    "extdata", "example_2.fasta", package = "tmhmm"
+  topology_text <- c(">A", "i", ">B", "ii")
+  expect_warning(
+    plot_locatome(topology_text),
+    "'plot_locatome' is deprecated, use 'plot_topology_text' instead"
   )
-  locatome <- run_tmhmm(fasta_filename)
-  expect_silent(plot_locatome(locatome))
 })