+++ NEWS.md
@@ -22,7 +22,9 @@ Newest versions at top.
 ### DEPRECATED AND DEFUNCT
 
   * `parse_fasta_file` is deprecated, 
-    use `pureseqtmr::load_fasta_file_as_tibble_cpp` instead
+    use `pureseqtmr::load_fasta_file_as_tibble` instead
+  * `locatome_to_df` is deprecated, use `fasta_text_to_tibble` instead
+  * Replace the argument name `locatome` by the more descriptive `topology_text`
 
 ## tmhmm 1.3 (2020-10-08)